Write a C program that prints the size of various types on the computer it is compiled and run on.

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ int main(void)
 	long int f;
 	long long int y;
 	float v;
-printf("size of a char: %zu byte(s)\n", (unsigned long)sizeof(x));
-printf("size of a int:  %zu byte(s)\n", (unsigned long)sizeof(z));
-printf("size of a long int: %zu byte(s)\n", (unsigned long)sizeof(f));
-printf("size of a long long int: %zu byte(s)\n", (unsigned long)sizeof(y));
-printf("size of a float: %zu byte(s)\n", (unsigned long)sizeof(v));
+printf("size of a char: %zu byte(s)\n", sizeof(x));
+printf("size of a int:  %zu byte(s)\n", sizeof(z));
+printf("size of a long int: %zu byte(s)\n", sizeof(f));
+printf("size of a long long int: %zu byte(s)\n", sizeof(y));
+printf("size of a float: %zu byte(s)\n", sizeof(v));
 return (0);
 }
